requirements.txt
fastapi==0.95.1
pydantic==1.10.2
motor==2.5.1
uvicorn==0.22.0
python-dotenv==1.0.0
